requirements.txt
Django>=4.2,<5.0
PyGithub>=2.1
python-dotenv>=1.0
gunicorn>=21.2
